Struct atomicbox::AtomicBox

pub struct AtomicBox<T> { /* fields omitted */ }

A type that holds a single Box<T> value and can be safely shared between threads.

Implementations

impl<T> AtomicBox<T>

pub fn new(value: Box<T>) -> AtomicBox<T>

Creates a new AtomicBox with the given value.

Examples

use atomicbox::AtomicBox;

let atomic_box = AtomicBox::new(Box::new(0));

pub fn swap(&self, other: Box<T>, order: Ordering) -> Box<T>

Atomically set this AtomicBox to other and return the previous value.

This does not allocate or free memory, and it neither clones nor drops any values. other is moved into self; the value previously in self is returned.

ordering must be either Ordering::AcqRel or Ordering::SeqCst, as other values would not be safe if T contains any data.

Panics

Panics if ordering is not one of the two allowed values.

Examples

use std::sync::atomic::Ordering;
use atomicbox::AtomicBox;

let atom = AtomicBox::new(Box::new("one"));
let prev_value = atom.swap(Box::new("two"), Ordering::AcqRel);
assert_eq!(*prev_value, "one");

pub fn store(&self, other: Box<T>, order: Ordering)

Atomically set this AtomicBox to other and drop its previous value.

The AtomicBox takes ownership of other.

ordering must be either Ordering::AcqRel or Ordering::SeqCst, as other values would not be safe if T contains any data.

Panics

Panics if ordering is not one of the two allowed values.

Examples

use std::sync::atomic::Ordering;
use atomicbox::AtomicBox;

let atom = AtomicBox::new(Box::new("one"));
atom.store(Box::new("two"), Ordering::AcqRel);
assert_eq!(atom.into_inner(), Box::new("two"));

pub fn swap_mut(&self, other: &mut Box<T>, order: Ordering)

Atomically swaps the contents of this AtomicBox with the contents of other.

This does not allocate or free memory, and it neither clones nor drops any values. The pointers in *other and self are simply exchanged.

ordering must be either Ordering::AcqRel or Ordering::SeqCst, as other values would not be safe if T contains any data.

Panics

Panics if ordering is not one of the two allowed values.

Examples

use std::sync::atomic::Ordering;
use atomicbox::AtomicBox;

let atom = AtomicBox::new(Box::new("one"));
let mut boxed = Box::new("two");
atom.swap_mut(&mut boxed, Ordering::AcqRel);
assert_eq!(*boxed, "one");

pub fn into_inner(self) -> Box<T>

Consume this AtomicBox, returning the last box value it contained.

Examples

use atomicbox::AtomicBox;

let atom = AtomicBox::new(Box::new("hello"));
assert_eq!(atom.into_inner(), Box::new("hello"));

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the contained value.

This is safe because it borrows the AtomicBox mutably, which ensures that no other threads can concurrently access either the atomic pointer field or the boxed data it points to.

pub fn load_handle(&self, order: Ordering) -> Handle<T>

Returns a handle that matches the current held box.

load_handle takes an Ordering argument which describes the memory ordering of this operation. Possible values are SeqCst, Acquire and Relaxed.

pub fn load_pointer(&self, order: Ordering) -> *mut T

Returns a pointer to the currently held box. Using the pointer is unsafe for all of the usual reasons; e.g., the box might have been dropped from the atomic by another thread by the time the pointer is dereferenced. load takes an Ordering argument which describes the memory ordering of this operation. Possible values are SeqCst, Acquire and Relaxed.

Examples

use atomicbox::AtomicBox;
use std::sync::atomic::Ordering;

let mut box1 = Box::new("goodbye");
let atom = AtomicBox::new(Box::new("hello"));
let ptr = &mut *box1 as *mut &str;

atom.store(box1, Ordering::SeqCst);
assert_eq!(atom.load_pointer(Ordering::Relaxed), ptr);

pub fn compare_exchange(
&self,
current: Handle<T>,
new: Box<T>,
success: Ordering,
failure: Ordering
) -> Result<Box<T>, (Handle<T>, Box<T>)>

Stores a box into the atomic if the value held by the atomic matches the given current handle. The return value is a result indicating whether the new box was written and containing the previous box. On success the returned handle is guaranteed to match the current value. On failure, the returned value contains the handle that matches the value in the atomic, and the given new box. compare_exchange takes two Ordering arguments to describe the memory ordering of this operation. success describes the required ordering for the read-modify-write operation that takes place if the comparison with current succeeds. failure describes the required ordering for the load operation that takes place when the comparison fails. Using Acquire as success ordering makes the store part of this operation Relaxed, and using Release makes the successful load Relaxed. The failure ordering can only be SeqCst, Acquire or Relaxed and must be equivalent to or weaker than the success ordering. Note: This method is only available on platforms that support atomic operations on pointers.

Examples

use atomicbox::AtomicBox;
use std::sync::atomic::Ordering;

let atom = AtomicBox::new(Box::new("hello"));
let box1 = Box::new("goodbye");
let current = atom.load_handle(Ordering::Relaxed);
let result = atom.compare_exchange(current, box1, Ordering::SeqCst, Ordering::Relaxed);
assert_eq!(result, Ok(Box::new("hello")));

pub fn compare_exchange_mut(
&self,
current: Handle<T>,
new: &mut Box<T>,
success: Ordering,
failure: Ordering
) -> Result<Handle<T>, Handle<T>>

Stores the new box into the atomic and moves the value in the atomic into the new box, if the value held by the atomic matches the given current handle. The return value is a result indicating whether the new box was written. On success the returned handle is guaranteed to match the current value. On failure, the returned value contains the handle that matches the value in the atomic, and new isn’t mutated. compare_exchange_mut takes two Ordering arguments to describe the memory ordering of this operation. success describes the required ordering for the read-modify-write operation that takes place if the comparison with current succeeds. failure describes the required ordering for the load operation that takes place when the comparison fails. Using Acquire as success ordering makes the store part of this operation Relaxed, and using Release makes the successful load Relaxed. The failure ordering can only be SeqCst, Acquire or Relaxed and must be equivalent to or weaker than the success ordering. Note: This method is only available on platforms that support atomic operations on pointers.

Examples

use atomicbox::AtomicBox;
use std::sync::atomic::Ordering;

let atom = AtomicBox::new(Box::new("hello"));
let mut box1 = Box::new("goodbye");
let current = atom.load_handle(Ordering::Relaxed);
atom.compare_exchange_mut(current, &mut box1, Ordering::SeqCst, Ordering::Relaxed);
assert_eq!(box1, Box::new("hello"));

pub fn compare_exchange_weak(
&self,
current: Handle<T>,
new: Box<T>,
success: Ordering,
failure: Ordering
) -> Result<Box<T>, (Handle<T>, Box<T>)>

Stores a box into the atomic if the value held by the atomic matches the given current handle. Unlike AtomicBox::compare_exchange, this function is allowed to spuriously fail even when the comparison succeeds, which can result in more efficient code on some platforms. The return value is a result indicating whether the new box was written and containing the previous box. On success the returned handle is guaranteed to match the current value. On failure, the returned value contains the handle that matches the value in the atomic, and the given new box unmutated. compare_exchange_weak takes two Ordering arguments to describe the memory ordering of this operation. success describes the required ordering for the read-modify-write operation that takes place if the comparison with current succeeds. failure describes the required ordering for the load operation that takes place when the comparison fails. Using Acquire as success ordering makes the store part of this operation Relaxed, and using Release makes the successful load Relaxed. The failure ordering can only be SeqCst, Acquire or Relaxed and must be equivalent to or weaker than the success ordering. Note: This method is only available on platforms that support atomic operations on pointers.

Examples

use atomicbox::AtomicBox;
use std::sync::atomic::Ordering;

let atom = AtomicBox::new(Box::new("hello"));
let mut box1 = Box::new("goodbye");
let mut current = atom.load_handle(Ordering::Relaxed);
box1 = loop {
    match atom.compare_exchange_weak(current, box1, Ordering::SeqCst, Ordering::Relaxed) {
        Ok(b) => break b,
        Err((c, b)) => {
            current = c;
            box1 = b;
        }
    }
};

pub fn compare_exchange_weak_mut(
&self,
current: Handle<T>,
new: &mut Box<T>,
success: Ordering,
failure: Ordering
) -> Result<Handle<T>, Handle<T>>

Moves the value in the new box into the atomic and move the value in the atomic into the new box, if the value held by the atomic matches the given current handle. Unlike AtomicBox::compare_exchange_mut, this function is allowed to spuriously fail even when the comparison succeeds, which can result in more efficient code on some platforms. The return value is a result indicating whether the new box was written. On success the returned handle is guaranteed to match the current value. On failure, the returned value contains the handle that matches the value in the atomic, and new isn’t mutated. compare_exchange_weak_mut takes two Ordering arguments to describe the memory ordering of this operation. success describes the required ordering for the read-modify-write operation that takes place if the comparison with current succeeds. failure describes the required ordering for the load operation that takes place when the comparison fails. Using Acquire as success ordering makes the store part of this operation Relaxed, and using Release makes the successful load Relaxed. The failure ordering can only be SeqCst, Acquire or Relaxed and must be equivalent to or weaker than the success ordering. Note: This method is only available on platforms that support atomic operations on pointers.

Examples

use atomicbox::AtomicBox;
use std::sync::atomic::Ordering;

let atom = AtomicBox::new(Box::new("hello"));
let mut box1 = Box::new("goodbye");
let mut current = atom.load_handle(Ordering::Relaxed);
loop {
    match atom.compare_exchange_weak_mut(current, &mut box1, Ordering::SeqCst, Ordering::Relaxed) {
        Ok(_) => break,
        Err(c) => current = c,
    }
}